refactor(ErrorMessage): ErrorMessage 통합 및 상수명 변경
