+++ src/main/java/ladder/model/Height.java
@@ -1,5 +1,7 @@
 package ladder.model;
 
+import static ladder.model.ErrorMessage.EXCEPTION_HEIGHT_MINIMUM;
+
 public class Height {
     private static final int MIN_HEIGHT = 2;
     private final int height;
@@ -11,25 +13,12 @@ public class Height {
 
     private void validateHeight(int height) {
         if (height < MIN_HEIGHT) {
-            throw new IllegalArgumentException(ErrorMessage.EXCEPTION_INVALID_HEIGHT.getMessage());
+            throw new IllegalArgumentException(EXCEPTION_HEIGHT_MINIMUM.getMessage());
         }
     }
 
     public int getHeight() {
         return height;
-    }
-
-    private enum ErrorMessage {
-        EXCEPTION_INVALID_HEIGHT("사다리 높이는 2 이상이어야 합니다.");
-        private final String message;
-
-        ErrorMessage(String message) {
-            this.message = message;
-        }
-
-        private String getMessage() {
-            return message;
-        }
     }
 
 }
