+++ src/main/java/ladder/model/PlayerName.java
@@ -1,5 +1,7 @@
 package ladder.model;
 
+import static ladder.model.ErrorMessage.EXCEPTION_PLAYER_NAME_LENGTH;
+
 public class PlayerName {
 
     private static final int MIN_LENGTH = 1;
@@ -18,25 +20,12 @@ public class PlayerName {
 
     private void validatePlayerNameLength(String playerName) {
         if (playerName.length() < MIN_LENGTH || playerName.length() > MAX_LENGTH) {
-            throw new IllegalArgumentException(ErrorMessage.EXCEPTION_INVALID_LENGTH_NAME.getMessage());
+            throw new IllegalArgumentException(EXCEPTION_PLAYER_NAME_LENGTH.getMessage());
         }
     }
 
     public String getPlayerName() {
         return playerName;
-    }
-
-    private enum ErrorMessage {
-        EXCEPTION_INVALID_LENGTH_NAME("플레이어 이름은 1자 이상 5자 이하여야 합니다.");
-        private final String message;
-
-        ErrorMessage(String message) {
-            this.message = message;
-        }
-
-        private String getMessage() {
-            return message;
-        }
     }
 
 }
